requirements.txt
aioredis==1.1.0
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.2
Automat==0.6.0
boto3==1.7.28
botocore==1.10.30
certifi==2018.4.16
cffi==1.11.5
channels==2.1.1
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
daphne==2.1.2
dj-database-url==0.5.0
Django==2.0.5
docutils==0.14
gunicorn==19.8.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jmespath==0.9.3
lxml==4.2.1
msgpack==0.5.6
msgpack-python==0.5.6
parsel==1.4.0
psycopg2==2.7.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-scrapyd-api==2.1.2
pytz==2018.4
pywin32==223
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
Scrapy==1.5.0
scrapy-heroku==0.7.1
scrapyd==1.2.0
scrapyd-client==1.2.0a1
scrapyd-heroku==0.1.0
selenium==3.12.0
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
uuid==1.30
w3lib==1.19.0
whitenoise==3.3.1
zope.interface==4.5.0
